feat: divide header in components
